pyproject.toml
[tool.poetry]
name = "shennongname"
version = "1.1.0"
description = "ShennongName (神农命名)"
authors = ["YANG Zijie <yangzijie@westlake.edu.cn>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"

pypinyin = "^0.49.0"
flask = "^3.0.0"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
pandas = "^2.1.1"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# If you want to publish the package, please:
# 1. run `poetry install` to install all dependencies
# 2. run `pytest` to make sure all tests pass
# 3. run `poetry build` to build the package
# 4. run `poetry publish` to publish the package